requirements.txt
# Flask и расширения
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1

# База данных
psycopg2-binary==2.9.7

# Excel
openpyxl==3.1.2

# Утилиты
python-dotenv==1.0.0
email-validator==2.0.0

# Development
pytest==7.4.2
pytest-cov==4.1.0